Apply mentions filter logic to comment text token
This begins by applying the mentions filter with Open Social's default
configuration to the comment text token. However, because e-mails are
created by cron which is executed as AN user, the query for profiles is
subject to access handling for AN users. This causes no profiles to be
loaded in the e-mail which prevents mention replacement from taking
place.

A workaround for this will be needed, either disabling access checking
for mentions (which could expose hidden entities if mentions is extended
beyond users). An alternative is to re-implement mentions as a data
transformation instead of a custom markup language, similarly to what
can be done for images with captions (see Drupal.org Filter
documentation). This could allow for a placeholder that is safe for AN
users so that a failing replacement doesn't matter.