Fix bugs and fix dashboard usage of content list block
The block wouldn't render on layout builder pages because the correct
hook was not implemented.

Additionally, implementing that hook exposed some bugs. When no fields
were selected, the `empty` value of the field selector would not
actually be empty but would be interpreted as a special field. This is
fixed by checking for that field name, although this will break if a
field named `all` is every added. So it's probably better to just make
the field be actually empty.

Additionally a query alter would fail if there were no filters selected
because it attempted to do a join with an `IN` clause without and
values. This is not allowed in a SQL database.