Provide social_search_comments module
This contains a configuration override that will add the comments for
nodes to the search indices and search views using the newly created
teaser view mode. Enabling the module and re-indexing after clearing the
cache should be enough.

If that's not enough you may need to change something insignificant in
an index and resave to retrigger the index caching as we've had to do
with overrides for search (e.g. SOLR or content tagging) previously.