Move pagination test logic to SocialGraphQLTestBase
This commit introduces a base class for Open Social GraphQL tests. This
class performs the set-up of the Open Social GraphQL server for testing.
It also introduces the `assertQuery` method which was used in the user
endpoint tests to verify queries produced the expected outcome.

Most importantly the test cases for pagination are generalised so that
they can easily be used in other endpoint tests. This makes testing
proper pagination support as easy as calling
`assertEndpointsupportsPagination('endpoint', $test_data)`.