Issue #3134488 by Taco Potze, collinm, ronaldtebrake, Kingdutch: Image with Post does not work after removing the image
It appears that an AJAX call introduces a random string after the
selectors that are used in the script. This breaks the script. We're
unable to trace the source of this change but we can make the script
more robust.

We only check for the start of data-selector. We also limit all queries
to the current concept in an attempt to make this work when there are
multiple forms on the same page.

The CSS is slightly updated because it appears that the line is no
longer needed but instead causes issues. The comment was correct that
the stylesheet was only loaded after the image is shown, but it's not
unloaded when the image is removed. This caused a visual bug that was
present until a very early (<1.7) version of Open Social.