Add resolver for DateTime type
This makes the results of timestamp fields available to clients. This is
used in quite a few places but wasn't actually retrievable by the
client.

For other formats it's probably easier to have a DateTime class that's
passed around so some investigation is needed on what we get from time
related fields and what best to pass around.