Move todos to Drupal.org and fix coding standards
This commit fixes coding standards in the social_graphql module and
moves todo's that require follow-up before a stable version to
Drupal.org

We exclude two rules of the coding standards for social_graphql as
violations are caused by the use of PHP 7.4 syntax for which Drupal
Coder is not yet ready.

Issue #3191651 has been filed with the module to address this.