Mark social_graphql module as hidden
The module is not feature complete. It's included for easier deployment
within Open Social that allows us to experiment with it and develop it
further. Classes and schema's in the module will be changed and broken
so enabling the module could break production environments. This is why
we don't expose the module.