Fix cursor in pageInfo for reversed lists
The GraphQL specification states that the startCursor and endCursor must
always correspond to the first and last items in the list. This is not
currently the case when a reversed result is requested or we're looking
at the last items in the list. We must make the logic for determining
the last cursor slightly more intelligent.

This commit also rewrites the way pagination is tested to ensure that
the tests actually fail for the code before this commit to demonstrate
that the bug is present (and prevent regression).

The changes to the test also ensures that it's no longer dependent on
the implementation details of the cursor. This means that the way
cursors are created can be changed without having to change the tests.
This is achieved by using the result of the first page in a request for
the second page. That proves that the returned cursor value is accurate.